Updated README.md to include list of systems demonstrated in this source code
Most of these systems do not have documentation or online resources for learning how to implement them, nor are there examples or guidelines for best-practice implementation. The listed systems of this project are designed to maximize their effectiveness and optimality, and thus demonstrate and can be used as guides for best-practice implementation of these systems.